Refactor commands to new file
